Create a GUID module
I didn't like the way I did GUID constants in Janitor, I need to work with a Windows API that needs a unique ID, and we need a GUID for tracking installation lifetimes. With all that in mind it made sense to go ahead and build a GUID generator/parsing module that can create CLSID handles.